Attention decoder not resolved

+++ model.py
@@ -127,7 +127,7 @@ class SummarizationModel(object):
                  in range(hps.num_word_dependency_labels)} for _ in range(hps.batch_size)]
             self._word_gcn_dropout = tf.placeholder_with_default(hps.word_gcn_dropout, shape=(), name='dropout')
             self._max_word_seq_len = tf.placeholder(tf.int32, shape=(), name='max_word_seq_len')
-            self._word_neighbour_count = tf.placeholder(tf.int32, [hps.batch_size, None], name='word_neighbour_count')
+            self._word_neighbour_count = tf.placeholder(tf.float32, [hps.batch_size, None], name='word_neighbour_count')
 
         if FLAGS.query_gcn:
             self._query_adj_in = [
@@ -138,7 +138,7 @@ class SummarizationModel(object):
                  in range(hps.num_word_dependency_labels)} for _ in range(hps.batch_size)]
             self._query_gcn_dropout = tf.placeholder_with_default(hps.word_gcn_dropout, shape=(), name='query_dropout')
             self._max_query_seq_len = tf.placeholder(tf.int32, shape=(), name='max_query_seq_len')
-            self._query_neighbour_count = tf.placeholder(tf.int32, [hps.batch_size, None], name='query_neighbour_count')
+            self._query_neighbour_count = tf.placeholder(tf.float32, [hps.batch_size, None], name='query_neighbour_count')
